# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/drinks": {
      "get": {
        "description": "Combinable filters and cursor pagination for infinite scroll: pass `nextCursor` as `cursor` to get the next page. Sorted by name. Alcoholic drinks are only returned to authenticated adults.",
        "operationId": "DrinksController_explore_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "description": "Part of the name",
              "example": "daiquiri",
              "type": "string",
              "maxLength": 60
            },
            "description": "Part of the name"
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "description": "As in /v1/drinks/facets",
              "example": "Cocktail",
              "type": "string"
            },
            "description": "As in /v1/drinks/facets"
          },
          {
            "name": "glass",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Highball glass",
              "type": "string"
            }
          },
          {
            "name": "alcoholic",
            "required": false,
            "in": "query",
            "schema": {
              "description": "true / false / omit for any",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "true / false / omit for any"
          },
          {
            "name": "iba",
            "required": false,
            "in": "query",
            "schema": {
              "description": "Only (true) or never (false) IBA official drinks",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Only (true) or never (false) IBA official drinks"
          },
          {
            "name": "ingredients",
            "required": false,
            "in": "query",
            "schema": {
              "description": "Comma separated; the drink must have all of them. Spanish names work.",
              "example": "ron,limón",
              "default": "",
              "type": "string"
            },
            "description": "Comma separated; the drink must have all of them. Spanish names work."
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 24,
              "type": "integer",
              "minimum": 1,
              "maximum": 60
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "description": "Opaque. Use `nextCursor` from the previous page",
              "type": "string"
            },
            "description": "Opaque. Use `nextCursor` from the previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of drink cards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrinkPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Explore the catalog",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/facets": {
      "get": {
        "description": "Categories, glasses and most used ingredients with how many drinks each has (Spanish names included), to build the filters UI.",
        "operationId": "DrinksController_facets_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Facets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrinkFacets"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Filter options",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/suggest": {
      "get": {
        "description": "Drinks and ingredients matching what the user is typing, tolerant to typos (\"margarta\", \"mojto\"). Search-as-you-type friendly.",
        "operationId": "DrinksController_suggest_v1",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 60,
              "description": "What the user is typing. Typos are fine",
              "example": "margarta"
            },
            "description": "What the user is typing. Typos are fine"
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 8,
              "type": "integer",
              "minimum": 1,
              "maximum": 15
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions, best first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Suggestion"
                  },
                  "description": "Drinks and ingredients, best match first"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Autocomplete",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/search": {
      "get": {
        "description": "Searches TheCocktailDB and stores the results; falls back to the local catalog if it is down. Alcoholic drinks are only returned to authenticated adults.",
        "operationId": "DrinksController_search_v1",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Part of the drink name",
              "example": "margarita"
            },
            "description": "Part of the drink name"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching drinks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Drink"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Search drinks by name",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/random": {
      "get": {
        "description": "Alcoholic drinks are only returned to authenticated adults.",
        "operationId": "DrinksController_random_v1",
        "parameters": [
          {
            "name": "alcoholic",
            "required": false,
            "in": "query",
            "schema": {
              "description": "true: only alcoholic; false: only alcohol-free; omit: any",
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "true: only alcoholic; false: only alcohol-free; omit: any"
          }
        ],
        "responses": {
          "200": {
            "description": "A random drink",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drink"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: role, ownership or age restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "A random drink",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/of-the-day": {
      "get": {
        "description": "The same drink for everyone during a UTC day. Alcoholic drinks are only returned to authenticated adults.",
        "operationId": "DrinksController_ofTheDay_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Today's drink",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrinkOfTheDay"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (TheCocktailDB) is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Drink of the day",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/{id}": {
      "get": {
        "description": "Alcoholic drinks are only returned to authenticated adults.",
        "operationId": "DrinksController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "TheCocktailDB drink id",
              "example": "11007"
            },
            "description": "Drink id"
          }
        ],
        "responses": {
          "200": {
            "description": "The drink",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Drink"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: role, ownership or age restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Drink detail",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/{id}/dna": {
      "get": {
        "description": "Flavor profile (sweet, sour, bitter, strong, fruity, herbal, creamy, fizzy, spicy), dominant traits, personality, strength and complexity.",
        "operationId": "DrinksController_dna_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "TheCocktailDB drink id",
              "example": "11007"
            },
            "description": "Drink id"
          }
        ],
        "responses": {
          "200": {
            "description": "The flavor DNA",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlavorDna"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: role, ownership or age restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Flavor DNA",
        "tags": [
          "Drinks"
        ]
      }
    },
    "/v1/drinks/{id}/twins": {
      "get": {
        "description": "Most similar drinks: 60% shared ingredients (Jaccard) + 40% flavor similarity (cosine).",
        "operationId": "DrinksController_twins_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "TheCocktailDB drink id",
              "example": "11007"
            },
            "description": "Drink id"
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 5,
              "description": "How many twins to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            },
            "description": "How many twins to return"
          }
        ],
        "responses": {
          "200": {
            "description": "The twins, most similar first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrinkTwins"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: role, ownership or age restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Twin drinks",
        "tags": [
          "Drinks"
        ]
      }
    },
...
  },
  "info": {
    "title": "api-drinks",
    "description": "\nAPI de bebidas y cocteles sobre TheCocktailDB: despensa inteligente, cocteles por estado de ánimo,\nADN de sabor, gemelos y carta inteligente para bares.\n\n**Versión**: todas las rutas van bajo `/v1`, salvo `/health`, `/health/ready` y `/`.\n\n**Errores**: todos tienen la forma `{ statusCode, code, message, error, details? }`. El frontend decide por `code`\n(estable) y muestra su propio texto; `message` es solo una pista en inglés.\n\n**Autenticación**\n- Apps propias: `Authorization: Bearer <accessToken>` (ver /v1/auth/login). En navegador, el refresh token\n  viaja en una cookie httpOnly: haz las peticiones de /v1/auth con credentials: 'include'.\n- Integraciones de terceros: `X-API-Key: <key>` (ver /v1/me/api-keys). Cuentan contra la cuota diaria del plan\n  y responden con los headers `X-RateLimit-Limit`, `X-RateLimit-Remaining` y `X-RateLimit-Reset`.\n\n**Fuente de datos**: recetas e imágenes de [TheCocktailDB](https://www.thecocktaildb.com/). Todo frontend que\nmuestre bebidas debe mostrar \"Datos e imágenes: TheCocktailDB\" con un enlace a su sitio (condición de su uso).\n\n**Verificación de edad**: las bebidas con alcohol solo se muestran a usuarios autenticados mayores de 18 años.\n",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "DrinkPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DrinkCard"
            }
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Drinks matching the filters"
          },
          "nextCursor": {
            "nullable": true,
            "description": "null on the last page",
            "type": "string"
          }
        },
        "required": [
          "items",
          "total",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "DrinkOfTheDay": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-24",
            "description": "UTC day"
          },
          "drink": {
            "$ref": "#/components/schemas/Drink"
          }
        },
        "required": [
          "date",
          "drink"
        ],
        "additionalProperties": false
      },
...
    }
  }
}